requirements.txt
mattermostdriver>=7.3.1
requests>=2.28.0
python-dotenv>=1.0.0
